image/Building Footprint Extraction/requirements.txt
numpy
tqdm
torch
torchvision
Pillow